requirements.txt
httpx
beautifulsoup4
sentence-transformers
umap-learn
numpy
tqdm
scikit-learn
fastapi
uvicorn
watchfiles
